Refactor CLI error handling and return types for improved clarity
- Updated `Run` method in `CliApp` to return `CliExitCode` instead of `int`, enhancing type safety.
- Improved error handling by replacing generic exceptions with `UserException` for clearer feedback on user errors.
- Streamlined the `_Execute` method to return `CliExitCode` values, ensuring consistent exit codes across the application.
- Enhanced argument parsing in `CliCommandFactory` to throw `UserException` for invalid arguments, improving user experience.

+++ mfr8/Program.cs
@@ -1,4 +1,4 @@
 ﻿using Mfr8.Cli;
 
 var exitCode = CliApp.Run(args);
-Environment.Exit(exitCode);
+Environment.Exit((int)exitCode);
